+++ Code/Source/DioramaModuleInterface.cpp
@@ -12,6 +12,7 @@
 
 #include <Clients/Collider2DComponent.h>
 #include <Clients/Collision2DSystemComponent.h>
+#include <Clients/DioramaCamera2DComponent.h>
 #include <Clients/DioramaLightComponent.h>
 #include <Clients/DioramaSystemComponent.h>
 #include <Clients/SpriteComponent.h>
@@ -38,6 +39,7 @@ namespace Diorama
                 TilemapComponent::CreateDescriptor(),
                 Collider2DComponent::CreateDescriptor(),
                 DioramaLightComponent::CreateDescriptor(),
+                DioramaCamera2DComponent::CreateDescriptor(),
             });
     }