+++ Code/Source/Tools/DioramaEditorModule.cpp
@@ -10,6 +10,7 @@
 #include <Diorama/DioramaTypeIds.h>
 #include <DioramaModuleInterface.h>
 #include <Tools/EditorCollider2DComponent.h>
+#include <Tools/EditorDioramaCamera2DComponent.h>
 #include <Tools/EditorDioramaLightComponent.h>
 #include <Tools/EditorSpriteComponent.h>
 #include <Tools/EditorTilemapComponent.h>
@@ -36,6 +37,7 @@ namespace Diorama
                     EditorTilemapComponent::CreateDescriptor(),
                     EditorCollider2DComponent::CreateDescriptor(),
                     EditorDioramaLightComponent::CreateDescriptor(),
+                    EditorDioramaCamera2DComponent::CreateDescriptor(),
                 });
         }